+++ src/widget/image.rs
@@ -66,7 +66,7 @@ pub struct ImageWidget<R: gfx::Resources> {
 impl<R: gfx::Resources> widget::Widget<R> for ImageWidget<R> {
     fn render(
         &self, data: &mut render::RenderData<R>,
-        prev_area: &widget::RenderArea, offset: &mut widget::RenderOffset)
+        prev_area: &render::RenderArea, offset: &mut render::RenderOffset)
     {
         let pos = [prev_area.position[0] + offset.position[0], prev_area.position[1] + offset.position[1]];
         data.push_rect_textured(pos, self.size, self.texture.handle());
